refactor: split event_stream.rs into codec + transport modules
Convert event_stream.rs (1339 lines) into module directory:
- event_stream/codec.rs (546 lines) — wire format, EventFrame encode
- event_stream/mod.rs (815 lines) — transport state machine, I/O thread

Closes #366
